fixing traces

+++ include/LTL/SuccessorGeneration/ResumingSuccessorGenerator.h
@@ -40,7 +40,7 @@ namespace LTL {
                 return _last_state != NoLastState;
             }
 
-            size_t transition() const {
+            uint32_t transition() const {
                 return _tcounter - 1;
             }
